Add a structured interface to nlogger
This adds Structured and Entry as interfaces to `nlogger` and also makes
Structured the default implementation. There's a new interface called
Basic as well, but for now we're keeping Logger around just to not break
the world for everyone that updates just yet. However, the context
functions (recently introduced) have been changed to deal with
the Structured logger interface rather than Logger.

This also introduces a function called ToStructured() which takes a
logger that implements a Basic logger and wraps it into an
implementation of Structured. This allows people to still use a simple
logger with our libraries, but we can fully implement a structured
logger throughout the codebase.

The Logger interface is deprecated and will be removed.